\name{FPCA2D-package}
\alias{FPCA2D-package}
\alias{FPCA2D}
\docType{package}
\title{
  Two Dimensional Functional Principal Component Analysis
}
\description{
 Compute the two dimension functional principal component scores for a series of two dimension images.
}
\details{
\tabular{ll}{
Package: \tab FPCA2D\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-09-01\cr
License: \tab GPL-3\cr
}
}
\author{
Nan Lin,Momiao Xiong

Maintainer: Nan Lin <Nan.Lin@uth.tmc.edu>
}
\references{
Lin N, Jiang J, Guo S, Xiong M (2015) Functional Principal Component Analysis and Randomized Sparse Clustering Algorithm for Medical Image Analysis. PLoS ONE 10(7): e0132945. doi:10.1371/journal.pone.0132945
}
\keyword{ package }
